fix schema for update and insert columns in mutations
