+++ crates/connectors/ndc-postgres/src/schema/mutation/mod.rs
@@ -1,5 +1,6 @@
 //! Generate ndc-spec schema metadata for mutations.
 
+pub mod helpers;
 mod v1;
 mod v2;
