+++ crates/connectors/ndc-postgres/src/schema/mutation/v1.rs
@@ -7,6 +7,7 @@ use ndc_sdk::models;
 use query_engine_translation::translation::mutation;
 
 use super::super::helpers::*;
+use super::helpers::*;
 
 /// given a v1 `DeleteMutation`, turn it into a `ProcedureInfo` to be output in the schema
 pub fn delete_to_procedure(
